requirements.txt
alembic==1.15.2
altgraph==0.17.4
aniso8601==10.0.1
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
black==25.1.0
blinker==1.9.0
bs4==0.0.2
build==1.2.2.post1
CacheControl==0.14.3
certifi==2025.1.31
cffi==1.17.1
cfgv==3.4.0
charset-normalizer==3.4.1
cleo==2.1.0
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
coverage==7.8.1
crashtest==0.4.1
cycler==0.12.1
distlib==0.3.9
dnspython==2.7.0
dominate==2.9.1
dulwich==0.22.8
email_validator==2.2.0
fastjsonschema==2.21.1
ffmpeg==1.4
filelock==3.18.0
findpython==0.6.3
flake8==7.2.0
Flask==3.1.0
Flask-Bootstrap==3.3.7.1
Flask-JWT-Extended==4.7.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.57.0
fsspec==2025.5.1
google_images_download==2.8.0
greenlet==3.2.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.0
identify==2.6.12
idna==3.10
imageio==2.37.0
iniconfig==2.1.0
installer==0.7.0
isort==6.0.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.1.0
Jinja2==3.1.6
joblib==1.4.2
keyring==25.6.0
kiwisolver==1.4.8
labelImg==1.8.6
lazy_loader==0.4
lxml==5.3.2
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib==3.10.1
mccabe==0.7.0
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.0
mutagen==1.47.0
mypy==1.15.0
mypy_extensions==1.1.0
networkx==3.4.2
nodeenv==1.9.1
numpy==2.2.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.2.3
pathspec==0.12.1
pbs-installer==2025.6.10
pefile==2023.2.7
pillow==11.1.0
pkginfo==1.12.1.2
platformdirs==4.3.8
pluggy==1.6.0
poetry==2.1.3
poetry-core==2.1.3
pre_commit==4.2.0
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyflakes==3.3.2
pygame==2.6.1
pyinstaller==6.13.0
pyinstaller-hooks-contrib==2025.4
PyJWT==2.10.1
pylast==5.5.0
pyparsing==3.2.3
pyproject_hooks==1.2.0
PyQt5==5.15.11
PyQt5-Qt5==5.15.2
PyQt5_sip==12.17.0
PyQt6==6.9.0
PyQt6-Qt6==6.9.0
PyQt6_sip==13.10.0
PySocks==1.7.1
pytest==8.3.5
pytest-cov==6.1.1
pytest-flask==1.3.0
pytest-qt==4.4.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-telegram-bot==22.1
pytz==2025.2
pywin32-ctypes==0.2.3
PyYAML==6.0.2
qasync==0.27.1
RapidFuzz==3.13.0
regex==2024.11.6
requests==2.32.3
requests-mock==1.12.1
requests-toolbelt==1.0.0
ruff==0.11.10
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.31.0
setuptools==78.1.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
sympy==1.13.3
telegram==0.0.1
threadpoolctl==3.6.0
tifffile==2025.3.30
tokenizers==0.21.1
tomlkit==0.13.3
torch==2.7.1
torchaudio==2.7.1
tqdm==4.67.1
transformers==4.52.4
trio==0.29.0
trio-websocket==0.12.2
trove-classifiers==2025.5.9.12
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
virtualenv==20.31.2
visitor==0.1.3
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
WTForms==3.2.1
zstandard==0.23.0
